+++ csharp/Euler/Problems/Eight.cs
@@ -12,13 +12,13 @@ namespace Euler.Problems
 			var thousandDigits = Resources.ThousandDigitNumber;
 
 			// break up into 13 digit chunks
-			double max = 0;
+			ulong max = 0;
 			var indexCount = 0;
 
 			foreach (var digit in thousandDigits)
 			{
 				var chunk = thousandDigits.Skip(indexCount).Take(13).ToList();
-				double total = chunk.Select(x => double.Parse(x.ToString())).Multiply();
+				ulong total = chunk.Select(x => ulong.Parse(x.ToString())).Multiply();
 				max = total > max ? total : max;
 
 				indexCount++;
